Fix circular dep on list
